multicam-detection/backend/requirements.txt
opencv-python
websockets
ffmpeg-python
dlib
Pillow
numpy
requests